third_party/blink/web_tests/http/tests/security/mixedContent/insecure-audio-video-in-main-frame.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.setPopupBlockingEnabled(false);
}

window.addEventListener("message", function (e) {
    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a window that loads insecure HTML5 audio and video.  We should
trigger a mixed content callback because the main frame in the window is HTTPS but is
displaying insecure content.</p>
<script>
window.addEventListener('load', function () {
    window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html");
});
</script>
</body>
</html>
