third_party/blink/web_tests/http/tests/security/mixedContent/insecure-audio-video-in-main-frame-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html' was loaded over HTTPS, but requested an insecure audio file 'http://example.test:8080/resources/test.mp4'. This content should also be served over HTTPS.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-audio-video.html' was loaded over HTTPS, but requested an insecure video 'http://example.test:8080/resources/test.mp4'. This content should also be served over HTTPS.
This test opens a window that loads insecure HTML5 audio and video. We should trigger a mixed content callback because the main frame in the window is HTTPS but is displaying insecure content.
